Improve enum handling for OperKind / InoutKind.
This fixes part of the Windows build issues (those parts related
to these enums).

It also makes sure that the values for the enumerations are mapped
the same between the C and Rust code.

At issue is that in C, the underlying type of the enumeration
is not guaranteed to be any one type in a portable way and
Windows and the rest of the world disagree about these enums
being either `i32` on Windows or `u32` elsewhere. So, we'll treat
them as `u32` for the Rust side, and make sure to cast to the
underlying type (`ecs_inout_kind_t` / `ecs_oper_kind_t`) when
interacting with the C world.