`ObserverBuilder`: Simplify `event_count`
This is only used from Rust to track the current index into the
description's events for adding events. We can make it a `usize`
instead of an `i32` and casting it as we don't need this to be
the exact same as in C++.

Further, the `ObserverBuilder::event_count()` function is not needed
and can be removed. (This is a change to the public API.)